evo2-service/pyproject.toml
[project]
name = "evo2-service"
version = "1.0.0"
description = "Self-hosted DNA embedding service using the Evo2 foundation model"
requires-python = ">=3.12"
dependencies = [
    "evo2",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
    "numpy>=1.26",
]

[dependency-groups]
dev = ["pytest", "pytest-asyncio", "httpx"]
